Revert "uint64_t rather than unsigned."
This reverts commit 0228552be60e90a5d1757d6a2f6e389459b8bb58.

+++ libco/libco.h
@@ -14,7 +14,7 @@ extern "C" {
    /* Sets size of main stack and stack space for other threads in portable version.
       Defaults to 64K and 256K, respectively. Must be set before first call to
       co_create(), otherwise it has no effect. */
-   void co_setmax( uint64_t main_size, uint64_t total_stack );
+   void co_setmax( unsigned main_size, unsigned total_stack );
 
    /* Currently active thread. Before first call to co_switch(), returns main thread. */
    cothread_t co_active( void );
@@ -22,7 +22,7 @@ extern "C" {
    /* Creates new thread with stack of size bytes and executes entry() when next
       switched to. Returns pointer to thread, or NULL if it failed. User entry()
       function must NEVER return. */ 
-   cothread_t co_create( uint64_t size, void (*entry)( void ) );
+   cothread_t co_create( unsigned int size, void (*entry)( void ) );
 
    /* Switches to thread. Thread must not already be the active one. */
    void co_switch( cothread_t thread );
