+++ nall/sort.hpp
@@ -19,14 +19,14 @@
 
 namespace nall {
   template<typename T>
-  void sort(T list[], uint64_t length) {
+  void sort(T list[], unsigned length) {
     if(length <= 1) return;  //nothing to sort
 
     //use insertion sort to quickly sort smaller blocks
     if(length < 64) {
-      for(uint64_t i = 0; i < length; i++) {
-        uint64_t min = i;
-        for(uint64_t j = i + 1; j < length; j++) {
+      for(unsigned i = 0; i < length; i++) {
+        unsigned min = i;
+        for(unsigned j = i + 1; j < length; j++) {
           if(list[j] < list[min]) min = j;
         }
         if(min != i) swap(list[i], list[min]);
@@ -35,15 +35,15 @@ namespace nall {
     }
 
     //split list in half and recursively sort both
-    uint64_t middle = length / 2;
+    unsigned middle = length / 2;
     sort(list, middle);
     sort(list + middle, length - middle);
 
     //left and right are sorted here; perform merge sort
     T *buffer = new T[length];
-    uint64_t offset = 0;
-    uint64_t left   = 0;
-    uint64_t right  = middle;
+    unsigned offset = 0;
+    unsigned left   = 0;
+    unsigned right  = middle;
     while(left < middle && right < length) {
       if(list[left] < list[right]) {
         buffer[offset++] = list[left++];
@@ -54,7 +54,7 @@ namespace nall {
     while(left  < middle) buffer[offset++] = list[left++];
     while(right < length) buffer[offset++] = list[right++];
 
-    for(uint64_t i = 0; i < length; i++) list[i] = buffer[i];
+    for(unsigned i = 0; i < length; i++) list[i] = buffer[i];
     delete[] buffer;
   }
 }
