+++ nall/utf8.hpp
@@ -30,7 +30,7 @@ namespace nall {
 
     utf16_t(const char *s = "") {
       if(!s) s = "";
-      uint64_t length = MultiByteToWideChar(CP_UTF8, 0, s, -1, 0, 0);
+      unsigned length = MultiByteToWideChar(CP_UTF8, 0, s, -1, 0, 0);
       buffer = new wchar_t[length + 1]();
       MultiByteToWideChar(CP_UTF8, 0, s, -1, buffer, length);
     }
@@ -56,7 +56,7 @@ namespace nall {
 
     utf8_t(const wchar_t *s = L"") {
       if(!s) s = L"";
-      uint64_t length = WideCharToMultiByte(CP_UTF8, 0, s, -1, 0, 0, (const char*)0, (BOOL*)0);
+      unsigned length = WideCharToMultiByte(CP_UTF8, 0, s, -1, 0, 0, (const char*)0, (BOOL*)0);
       buffer = new char[length + 1]();
       WideCharToMultiByte(CP_UTF8, 0, s, -1, buffer, length, (const char*)0, (BOOL*)0);
     }
@@ -66,14 +66,14 @@ namespace nall {
     }
 
     utf8_t(const utf8_t & s){
-      uint64_t length = _mbslen((uint64_t char const *)s.buffer);
+      unsigned length = _mbslen((unsigned char const *)s.buffer);
       buffer = new char[length + 1]();
       memcpy(buffer, s.buffer, length + 1);
     }
 
     utf8_t& operator=(const utf8_t & s){
       delete [] buffer;
-      uint64_t length = _mbslen((uint64_t char const *)s.buffer);
+      unsigned length = _mbslen((unsigned char const *)s.buffer);
       buffer = new char[length + 1]();
       memcpy(buffer, s.buffer, length + 1);
     }
@@ -84,7 +84,7 @@ namespace nall {
   inline void utf8_args(int &argc, char **&argv) {
     wchar_t **wargv = CommandLineToArgvW(GetCommandLineW(), &argc);
     argv = new char*[argc];
-    for(uint64_t i = 0; i < argc; i++) {
+    for(unsigned i = 0; i < argc; i++) {
       argv[i] = new char[_MAX_PATH];
       strcpy(argv[i], nall::utf8_t(wargv[i]));
     }
